// src/app/shared/models/est_spat.model.ts
export class EstSpat {
  id?: number;
  no_spat?: string;
  jenis?:string;
  rayon_id?: number;
  is_double_handling?: boolean;
  divisi_id?: number;
  tanggal?: string;
  keterangan?: string;
  pks_timbangan_id:number;
  no_tiket:string;
  berat_bersih:number;
  total_jjg :number;
  total_brondolan :number;
  total_kg_kebun:number;
  detail?:[];

}